@@ -9,30 +9,31 @@
     "type": "project",
     "require": {
         "php": "^8.2",
-        "backpack/crud": "^5.0",
-        "backpack/revise-operation": "^1.0",
+        "backpack/crud": "^6.0",
+        "backpack/pro": "*",
+        "backpack/theme-coreuiv2": "^1.2",
+        "backpack/revise-operation": "^2.0",
         "box/spout": "^3.0",
-        "doctrine/dbal": "^3.1",
+        "doctrine/dbal": "^3.0",
         "spatie/ignition": "^1.0",
-        "fideloper/proxy": "~4.0",
         "lab404/laravel-impersonate": "^1.3",
-        "laravel/framework": "^9.0",
-        "laravel/helpers": "^1.2",
+        "laravel/framework": "^10.0",
+        "laravel/helpers": "^1.3",
         "laravel/legacy-factories": "^1.1",
-        "laravel/passport": "^10.0",
-        "laravel/sanctum": "^2.0",
+        "laravel/passport": "^11.0",
+        "laravel/sanctum": "^3.2",
         "laravel/slack-notification-channel": "^2.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^3.0",
+        "laravel/ui": "^4.0",
         "lorisleiva/laravel-actions": "^2.4",
         "predis/predis": "^1.1",
         "rap2hpoutre/laravel-log-viewer": "^2.0",
         "spatie/laravel-permission": "^6.0",
-        "venturecraft/revisionable": "^1.38",
+        "venturecraft/revisionable": "^1.0",
         "zircote/swagger-php": "^4.4"
     },
     "require-dev": {
-        "backpack/generators": "^3.0",
+        "backpack/generators": "^4.0",
         "barryvdh/laravel-debugbar": "^3.2",
         "fakerphp/faker": "^1.9.1",
         "mockery/mockery": "~1.0",
@@ -87,14 +88,15 @@
         "post-autoload-dump": [
             "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
             "@php artisan package:discover"
-        ]
+        ],
+        "post-install-cmd": ["php artisan storage:link --quiet"]
     },
     "config": {
         "preferred-install": "dist",
         "sort-packages": true,
         "optimize-autoloader": true
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "repositories": {
         "backpack": {
