@@ -6,8 +6,8 @@ use Illuminate\Database\Eloquent\Relations\Pivot;
 
 class CurationExpertPanel extends Pivot
 {
-    protected $dates = [
-        'start_date',
-        'end_date'
+    protected $casts = [
+        'start_date' => 'datetime',
+        'end_date' => 'datetime',
     ];
 }
