@@ -56,7 +56,7 @@ class AddStatus implements ShouldQueue
                 ],
             ]);
 
-            UpdateCurrentStatus::dispatchNow($this->curation);
+            UpdateCurrentStatus::dispatchSync($this->curation);
         });
     }
 
