@@ -34,7 +34,7 @@ class AugmentWithHgncInfo
             || is_null($event->curation->hgnc_name)
         ) {
             try {
-                HgncInfoJob::dispatchNow($event->curation);
+                HgncInfoJob::dispatchSync($event->curation);
             } catch (HttpNotFoundException $e) {
                 \Log::warning($e->getMessage());
             }
