@@ -29,6 +29,6 @@ class MakeCurationDeletedStreamMessage
      */
     public function handle(Deleted $event)
     {
-        \Bus::dispatchNow(new CreatePrecurationStreamMessage($event->curation, 'deleted'));
+        \Bus::dispatchSync(new CreatePrecurationStreamMessage($event->curation, 'deleted'));
     }
 }
