@@ -2,9 +2,11 @@
 
 namespace App\Logging;
 
+use Monolog\LogRecord;
+
 class ContainerRoleProcessor
 {
-    public function __invoke(array $record)
+    public function __invoke(LogRecord $record)
     {
         if (config('app.container_role')) {
             $record['channel'] = $record['channel'].':'.config('app.container_role');
